requirements.txt
fastapi[standard]
python-multipart
python-dotenv
flake8
Authlib==1.2.1
itsdangerous==2.1.2
pytest